feat(api): Separate prompt generation from API client
Refactors the `GeminiClient` to delegate prompt construction to a new, dedicated module, improving separation of concerns.

This addresses the feedback that the `generateContent` method was overly complex by handling two distinct jobs: provider-independent string forming and provider-dependent payload forming.

- A new `promptBuilder.js` module is introduced to handle all complex prompt engineering logic.
- The `GeminiClient` is now streamlined to focus solely on interacting with the Gemini API (payload creation, network requests, and response handling).

This change enhances code readability, testability, and maintainability by better adhering to the single-responsibility principle.
